+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fix setting the `stac` property for children
+- Allow to set specific options for children
 
 ## [1.0.0-rc.1] - 2025-01-25